move socket code at of api start
