+++ packages/server/src/users/api.ts
@@ -1,25 +1,12 @@
-import { uuid } from '@botpress/messaging-base'
 import { Router } from 'express'
 import { Auth } from '../base/auth/auth'
-import { ClientService } from '../clients/service'
-import { SocketManager } from '../socket/manager'
-import { SocketService } from '../socket/service'
-import { UserTokenService } from '../user-tokens/service'
-import { AuthUserSocketSchema, GetUserSchema, GetUserSocketSchema } from './schema'
+import { GetUserSchema } from './schema'
 import { UserService } from './service'
 
 export class UserApi {
-  constructor(
-    private router: Router,
-    private auth: Auth,
-    private clients: ClientService,
-    private sockets: SocketManager,
-    private users: UserService,
-    private userTokens: UserTokenService,
-    private socketService: SocketService
-  ) {}
+  constructor(private router: Router, private auth: Auth, private users: UserService) {}
 
-  async setup() {
+  setup() {
     this.router.post(
       '/users',
       this.auth.client.auth(async (req, res) => {
@@ -48,50 +35,6 @@ export class UserApi {
 
         res.send(user)
       })
-    )
-
-    this.sockets.handle('users.get', GetUserSocketSchema, async (socket) => {
-      socket.reply(await this.users.get(socket.userId))
-    })
-
-    // TODO: this should be done when establishing the socket connection
-    this.sockets.handle(
-      'users.auth',
-      AuthUserSocketSchema,
-      async (socket) => {
-        const { clientId, id: userId, token: userTokenRaw }: { clientId: uuid; id: uuid; token: string } = socket.data
-
-        const client = await this.clients.getById(clientId)
-        if (!client) {
-          return socket.notFound('Client not found')
-        }
-
-        // TODO: refactor here
-
-        let success = true
-        let user = userId ? await this.users.get(userId) : undefined
-        let token = undefined
-
-        if (!user || user.clientId !== client.id) {
-          success = false
-        } else {
-          const [userTokenId, userTokenToken] = userTokenRaw.split('.')
-          if (!(await this.userTokens.verifyToken(userTokenId, userTokenToken))) {
-            success = false
-          }
-        }
-
-        if (!success) {
-          user = await this.users.create(clientId)
-          const tokenRaw = await this.userTokens.generateToken()
-          const userToken = await this.userTokens.create(user.id, tokenRaw, undefined)
-          token = `${userToken.id}.${tokenRaw}`
-        }
-
-        this.socketService.registerForUser(socket.socket, user!.id)
-        socket.reply({ id: user!.id, token })
-      },
-      false
     )
   }
 }
